initial commit
"kompozitor" is the Russian word for "Composer".  This file acts like the
composer/conductor for the performance files it adds to the Chuck Virtual
Machine.  Prior to adding performance files it adds files from the utils
directory.

IMPORTANT: The module for any external class/function referenced in any
script MUST be added to VM prior to the script that requires it.

TODO: Figure out how to provide performance files as arguments